refactor(gctx): extract error to a module
